% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{Case_11}
\alias{Case_11}
\title{Gets the simulated data from Case_11}
\usage{
Case_11(x, n)
}
\arguments{
\item{x}{The x used to generate the clusters}

\item{n}{The number of curves per cluster}
}
\value{
The simulated data from Case_11
}
\description{
Gets the simulated data from Case_11
}
\examples{
Case_11(x, n)
}
