% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{Case_44}
\alias{Case_44}
\title{Gets the simulated data from Case_44}
\usage{
Case_44(x, n)
}
\arguments{
\item{x}{The x used to generate the clusters}

\item{n}{The number of curves per cluster}
}
\value{
The simulated data from Case_44
}
\description{
Gets the simulated data from Case_44
}
\examples{
Case_44(x, n)
}
