% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{Case_7}
\alias{Case_7}
\title{Gets simulated data from Case_7}
\usage{
Case_7(x, n)
}
\arguments{
\item{x}{The x used to generate the clusters}

\item{n}{The number of curves per cluster}
}
\value{
The simulated data from Case_7
}
\description{
Gets simulated data from Case_7
}
\examples{
Case_7(x, n)
}
