% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo_convergence.R
\name{check_convergence}
\alias{check_convergence}
\title{Checks if the algorithim has converged with the given threshold}
\usage{
check_convergence(prev_elbo, curr_elbo, convergence_threshold)
}
\arguments{
\item{prev_elbo}{The elbo value from the previous iteration}

\item{curr_elbo}{The elbo value from the current iteration}

\item{convergence_threshold}{The threshold that determines when the model has converged}
}
\value{
A boolean whether the algorithim has converged
}
\description{
Checks if the algorithim has converged with the given threshold
}
\examples{
check_convergence(prev_elbo, curr_elbo, convergence_threshold)
}
