% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.R
\name{compute_function}
\alias{compute_function}
\title{Runs and evaluates the model on the specified simulated function data for a single iteration}
\usage{
compute_function(
  simulation_number,
  case,
  x,
  K,
  curves_per_cluster,
  initialization_method,
  nbasis,
  true_cluster_assignments,
  gamma_dist_config_matrix,
  convergence_threshold,
  verbose,
  draw
)
}
\arguments{
\item{simulation_number}{The iteration number in the set of simulations}

\item{case}{The case used to simulate the data}

\item{x}{The dependent variable}

\item{K}{The number of clusters in the data}

\item{curves_per_cluster}{The number of curves per cluster}

\item{initialization_method}{The initialization method for the algorithim}

\item{nbasis}{The number of basis functions}

\item{true_cluster_assignments}{The true cluster assignments}

\item{gamma_dist_config_matrix}{A matrix where the rows are the alpha and parameters for each cluster}

\item{convergence_threshold}{The threshold that determines when the model has converged}

\item{verbose}{A boolean indicating whether or not to print the inner parameters of the model}

\item{draw}{A boolean indicating whether or not to draw the true function vs the}
}
\value{
A list with the various accuracy scores at each index
}
\description{
Runs and evaluates the model on the specified simulated function data for a single iteration
}
\examples{
compute_function(simulation_number, case, x, K, curves_per_cluster, initialization_method, nbasis, true_cluster_assignments, gamma_dist_config_matrix, convergence_threshold, verbose, draw)
}
