% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fit}
\alias{fit}
\title{Fit the model to the data}
\usage{
fit(
  Y,
  K,
  nbasis,
  x,
  init,
  true_cluster_assignments,
  gamma_dist_config_matrix,
  convergence_threshold
)
}
\arguments{
\item{Y}{The matrix containing rows corresponding the curves}

\item{K}{The number of clusters in the data}

\item{nbasis}{The number of basis functions}

\item{x}{The x used to generate the clusters}

\item{function_data}{The data generated from the actual functions}
}
\value{
The fitted model
}
\description{
File Wide Imports
}
\examples{
fit(Y, K)
}
