% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{funcslustVI}
\alias{funcslustVI}
\title{Generates cluster assignments and related information given functional data.}
\usage{
funcslustVI(
  x,
  Y,
  K,
  true_cluster_assignments,
  init,
  nbasis,
  convergence_threshold,
  max_iterations,
  gamma_dist_config_matrix,
  verbose,
  draw,
  plot_params
)
}
\arguments{
\item{x}{The dependent variable}

\item{Y}{A matrix in which the rows represent the curves}

\item{K}{The number of clusters in the data}

\item{true_cluster_assignments}{The true cluster assignments}

\item{init}{The initialization method for the algorithim}

\item{nbasis}{The number of basis functions}

\item{convergence_threshold}{The threshold that determines when the model has converged}

\item{max_iterations}{The maximum amount of iterations for the algorithim}

\item{gamma_dist_config_matrix}{A matrix where the rows are the alpha and parameters for each cluster}

\item{verbose}{A boolean indicating whether or not to print the inner parameters of the model}

\item{draw}{Parameter that determines whether or not to plot the real vs estimated curves after fitting the model}

\item{plot_params}{List of parameters corresponding to characteristics of the plot. Must include vectors xlim and ylim corresponding to the x and y limits of the plot.}
}
\value{
A list with entries containing varius information about the fitted model
}
\description{
Generates cluster assignments and related information given functional data.
}
\examples{
fit(x, Y, K, true_cluster_assignments, init, nbasis, convergence_threshold, gamma_dist_config_matrix, verbose, draw, plot_params)
}
