% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_A_vector}
\alias{get_A_vector}
\title{Generates a vector A}
\usage{
get_A_vector(alpha_vector, observations_per_curve)
}
\arguments{
\item{alpha_vector}{A vector that in which the entries are the alpha parameters of the gamma distribution (1 / variance) of the curves in each cluster}

\item{observations_per_curve}{The number of entries in each curve vector}
}
\value{
A vector A
}
\description{
Generates a vector A
}
\examples{
get_A(alpha_vector)
}
