% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_approx_function_data}
\alias{get_approx_function_data}
\title{Gets functional data approximations}
\usage{
get_approx_function_data(Y, K, probability_matrix)
}
\arguments{
\item{K}{The total number of clusters}

\item{probability_matrix}{A matrix in which the rows represent the probabilities that the curve is in each of the clusters}
}
\value{
Matrix with approximations of functional data for each cluster in a row
}
\description{
Gets functional data approximations
}
\examples{
get_approx_function_data(Y, number_of_cluster)
}
