% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_approx_phi_matrix}
\alias{get_approx_phi_matrix}
\title{Gets matrix of the coffecient vectors for each cluster, phi_k, of the basis matrix B}
\usage{
get_approx_phi_matrix(Y, K, nbasis, probability_matrix, x)
}
\arguments{
\item{Y}{The matrix containing rows corresponding the curves}

\item{K}{The total number of clusters}

\item{nbasis}{The number of basis functions}

\item{probability_matrix}{A matrix in which the rows represent the probabilities that the curve is in each of the clusters}

\item{x}{The x used to generate the clusters}
}
\value{
A matrix of the coffecient vectors for each cluster, phi_k, of the basis matrix B
}
\description{
Gets matrix of the coffecient vectors for each cluster, phi_k, of the basis matrix B
}
\examples{
get_approx_phi_matrix(Y, K, x, nbasis)
}
