% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_approx_probability_matrix_hcl}
\alias{get_approx_probability_matrix_hcl}
\title{Generates probability matrix with approximated probability values for each cluster}
\usage{
get_approx_probability_matrix_hcl(Y, K, x)
}
\arguments{
\item{K}{The number of clusters in cluster data}

\item{curves_per_cluster}{The number of curves per cluster}
}
\value{
A probability matrix with probability values for each cluster
}
\description{
Generates probability matrix with approximated probability values for each cluster
}
\examples{
get_equal_probability_matrix(K, curves_per_cluster)
}
