% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_approx_probability_matrix_km}
\alias{get_approx_probability_matrix_km}
\title{Generates probability matrix with approximated probability values for each cluster}
\usage{
get_approx_probability_matrix_km(Y, K, x)
}
\arguments{
\item{Y}{The matrix containing rows corresponding the curves}

\item{K}{The number of clusters in cluster data}

\item{x}{The dependent variable}
}
\value{
A probability matrix with approximated probability values for each cluster suign kmeans
}
\description{
Generates probability matrix with approximated probability values for each cluster
}
\examples{
get_approx_probability_matrix(Y, K, x)
}
