% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_B}
\alias{get_B}
\title{Generates B matrix}
\usage{
get_B(x, nbasis)
}
\arguments{
\item{x}{The x used to generate the clusters}

\item{nbasis}{The number of basis functions}

\item{total_number_of_curves}{The total number of curves}
}
\value{
A matrix in which row represent curves of various clusters
}
\description{
Generates B matrix
}
\examples{
get_B(x, total_number_of_curves, nbasis)
}
