% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_beta_vector}
\alias{get_beta_vector}
\title{Generates a vector in which the entries are the beta parameters of the gamma distribution (1 / variance) of the curves in each cluster}
\usage{
get_beta_vector(tau_list)
}
\arguments{
\item{tau_list}{A list in which each index is the tau vector for that cluster}
}
\value{
A vector that in which the entries are the beta parameters of the gamma distribution (1 / variance) of the curves in each cluster
}
\description{
Generates a vector in which the entries are the beta parameters of the gamma distribution (1 / variance) of the curves in each cluster
}
\examples{
get_beta_vector(tau_list, observations_per_cluster_x)
}
