% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_cumulative_phi_matrix}
\alias{get_cumulative_phi_matrix}
\title{Gets a matrix of the coffecient vectors, phi, for each curve}
\usage{
get_cumulative_phi_matrix(Y, x, nbasis)
}
\arguments{
\item{Y}{The matrix containing rows corresponding the curves}

\item{x}{The x used to generate the clusters}

\item{nbasis}{The number of basis functions}
}
\value{
a matrix of the coffecient vectors, phi, for each curve
}
\description{
Gets a matrix of the coffecient vectors, phi, for each curve
}
\examples{
get_cumulative_phi_matrix(Y, x, nbasis)
}
