% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_d_not_vector}
\alias{get_d_not_vector}
\title{Gets the d_not parameter for each cluster}
\usage{
get_d_not_vector(K)
}
\arguments{
\item{K}{The total number of clusters}
}
\value{
The d_not parameter vector for each cluster
}
\description{
Gets the d_not parameter for each cluster
}
\examples{
get_d_not_vector(K)
}
