% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo_convergence.R
\name{get_elbo}
\alias{get_elbo}
\title{Gets the elbo value for a single iteration of the algorithim}
\usage{
get_elbo(
  x,
  Y,
  K,
  phi_matrix,
  m_not_vector,
  nbasis,
  sigma_list,
  m_list,
  A_vector,
  R_vector,
  d_vector,
  probability_matrix,
  alpha_vector,
  beta_vector
)
}
\arguments{
\item{x}{The x used to generate the clusters}

\item{Y}{The matrix containing rows corresponding the curves}

\item{K}{The number of clusters in the data}

\item{phi_matrix}{A matrix of the coffecient vectors for each cluster, phi_k, of the basis matrix B}

\item{m_not_vector}{The vector containing m_not values for each cluster}

\item{nbasis}{The number of basis functions}

\item{m_list}{A list of the m parameters for each cluster}

\item{A_vector}{A vector}

\item{R_vector}{A vector}

\item{d_vector}{A vector of the d parameters for each cluster}

\item{probability_matrix}{A matrix in which the rows represent the probabilities that the curve is in each of the clusters}

\item{alpha_vector}{A vector that in which the entries are the alpha parameters of the gamma distribution (1 / variance) of the curves in each cluster}

\item{beta_vector}{A vector contain the beta vector for each cluster}
}
\value{
The elbo value
}
\description{
Gets the elbo value for a single iteration of the algorithim
}
\examples{
get_elbo(x, Y, K, phi_matrix, m_not_vector, nbasis, sigma_list, m_list, A_vector, R_vector, d_vector, probability_matrix, alpha_vector, beta_vector)
}
