% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_equal_probability_matrix}
\alias{get_equal_probability_matrix}
\title{Generates probability matrix with equal proability values for each cluster}
\usage{
get_equal_probability_matrix(cluster_data, number_of_clusters)
}
\arguments{
\item{number_of_clusters}{The number of clusters in cluster data}

\item{curves_per_cluster}{The number of curves per cluster}
}
\value{
A probability matrix with equal probability vlaues for each cluster
}
\description{
Generates probability matrix with equal proability values for each cluster
}
\examples{
get_equal_probability_matrix(number_of_clusters, curves_per_cluster)
}
