% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.R
\name{get_fdcvi_cluster_assignments}
\alias{get_fdcvi_cluster_assignments}
\title{Gets the clustering assignments generated by the algorithim}
\usage{
get_fdcvi_cluster_assignments(
  Y,
  K,
  nbasis,
  x,
  initialization_method,
  true_cluster_assignments,
  gamma_dist_config_matrix,
  convergence_threshold,
  verbose,
  draw
)
}
\arguments{
\item{Y}{A matrix in which the rows represent the curves}

\item{K}{The number of clusters in the data}

\item{nbasis}{The number of basis functions}

\item{x}{The dependent variable}

\item{initialization_method}{The initialization method for the algorithim}

\item{true_cluster_assignments}{The true cluster assignments}

\item{gamma_dist_config_matrix}{A matrix where the rows are the alpha and parameters for each cluster}

\item{convergence_threshold}{The threshold that determines when the model has converged}

\item{curves_per_cluster}{The number of curves per cluster}
}
\value{
The cluster assignments generated by the algorithim
}
\description{
Gets the clustering assignments generated by the algorithim
}
\examples{
get_fdcvi_cluster_assignments(Y, K, nbasis, x, initialization_method, true_cluster_assignments, gamma_dist_config_matrix, convergence_threshold, verbose, draw) 
}
