% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.R
\name{get_funclustVI_cluster_assignments}
\alias{get_funclustVI_cluster_assignments}
\title{Model function wrapper for the funclustVI}
\usage{
get_funclustVI_cluster_assignments(Y, data_params, model_params)
}
\arguments{
\item{Y}{A matrix in which the rows represent the curves}

\item{data_params}{List object containing the parameters required for generating the functional data and its characterics}

\item{model_params}{List object containing the parameters required for model the data and generating the cluster assignments}
}
\value{
The cluster assignments generated with the funclustVI model generated with the specified parameters
}
\description{
Model function wrapper for the funclustVI
}
\examples{
compute_function(seed, data_params, model_params, eval_func_list)
}
