% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.R
\name{get_mismatches}
\alias{get_mismatches}
\title{Gets the number of mismatches for a single iteration of the simulation}
\usage{
get_mismatches(cluster_assignments, K, curves_per_cluster)
}
\arguments{
\item{cluster_assignments}{A vector where each entry is the cluster assignment for the corresponging curve}

\item{K}{The number of clusters in the data}

\item{curves_per_cluster}{The number of curves per cluster}
}
\value{
The number of mismatches
}
\description{
Gets the number of mismatches for a single iteration of the simulation
}
\examples{
get_mismatches(cluster_assignments, K, curves_per_cluster)
}
