% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_tau_list}
\alias{get_tau_list}
\title{Generates a matrix in which each row represents the variances of the curves in each cluster}
\usage{
get_tau_list(Y, probability_matrix, K)
}
\arguments{
\item{Y}{The matrix containing rows corresponding to the curves}

\item{probability_matrix}{A matrix in which the rows represent the probabilities that the curve is in each of the clusters}

\item{K}{The number of entries in each curve vector}
}
\value{
A matrix in which each row represents the variances of the curves in each cluster
}
\description{
Generates a matrix in which each row represents the variances of the curves in each cluster
}
\examples{
get_tau_list(Y, probability_matrix, K)
}
