% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_true_m_not}
\alias{get_true_m_not}
\title{Gets the vector with the true m_not values}
\usage{
get_true_m_not(x, Y, K, nbasis, true_cluster_assignments)
}
\arguments{
\item{x}{The x used to generate the clusters}

\item{Y}{The matrix containing rows corresponding the curves}

\item{K}{The number of clusters in the data}

\item{nbasis}{The number of basis functions}

\item{true_cluster_assignments}{A vector containing the true cluster assignments}
}
\value{
The vector with the true m_not values
}
\description{
Gets the vector with the true m_not values
}
\examples{
get_true_m_no(x, Y, K, nbasis, true_cluster_assignments)
}
