% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_true_probability_matrix}
\alias{get_true_probability_matrix}
\title{Gets the true probability natrix using the true cluster assignments}
\usage{
get_true_probability_matrix(true_cluster_assignments, K)
}
\arguments{
\item{true_cluster_assignments}{A vector containing the true cluster assignments}

\item{K}{The number of clusters in the data}
}
\value{
The true probability matrix of the true cluster assignments
}
\description{
Gets the true probability natrix using the true cluster assignments
}
\examples{
get_true_probability_matrix(true_cluster_assignments, K)
}
