% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{Mismatch}
\alias{Mismatch}
\title{Gets a matrix of the final cluster assignments based on the probability matrix}
\usage{
Mismatch(predicited_clusters, true_clusters, K)
}
\arguments{
\item{predicited_clusters}{The predicted clusters}

\item{true_clusters}{THe true clusters}

\item{K}{The number of clusters}
}
\value{
a matrix of the final cluster assignments
}
\description{
Gets a matrix of the final cluster assignments based on the probability matrix
}
\examples{
Mismatch(predicited_clusters, true_clusters, K)
}
