% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.R
\name{simulate}
\alias{simulate}
\title{Runs and evaluates the model on the specified simulated function data}
\usage{
simulate(
  data_params,
  model_params,
  eval_func_list,
  number_of_simulations,
  save_path
)
}
\arguments{
\item{data_params}{List object containing the parameters required for generating the functional data and its characterics}

\item{model_params}{List object containing the parameters required for model the data and generating the cluster assignments}

\item{eval_func_list}{List object containing the functions corresponding to the various evaluations metrics}

\item{number_of_simulations}{The number of simulations}

\item{save_path}{The file path to save the results from the simulations}
}
\value{
A matrix with the results of the algorithim for iteration in the simulation for each metric being evaluated over
}
\description{
Runs and evaluates the model on the specified simulated function data
}
\examples{
simulate(data_params, model_params, eval_func_list, number_of_simulations, save_path)
}
