% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{update_d_vector}
\alias{update_d_vector}
\title{Update the d parameter for each cluster}
\usage{
update_d_vector(Y, K, probability_matrix)
}
\arguments{
\item{Y}{The matrix containing rows corresponding the curves}

\item{K}{The total number of clusters}

\item{probability_matrix}{A matrix in which the rows represent the probabilities that the curve is in each of the clusters}
}
\value{
A vector of the updated d parameters for each cluster
}
\description{
Update the d parameter for each cluster
}
\examples{
update_d_vector(Y, K, probability_matrix)
}
