% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{update_probability_matrix}
\alias{update_probability_matrix}
\title{Update the probabilty matrix}
\usage{
update_probability_matrix(
  Y,
  K,
  probability_matrix,
  x,
  sigma_list,
  m_list,
  A_vector,
  R_vector,
  d_vector,
  nbasis
)
}
\arguments{
\item{Y}{The matrix containing rows corresponding the curves}

\item{K}{The total number of clusters}

\item{probability_matrix}{A matrix in which the rows represent the probabilities that the curve is in each of the clusters}

\item{x}{The x used to generate the clusters}

\item{m_list}{A list of the m parameters for each cluster}

\item{A_vector}{A vector}

\item{R_vector}{A vector}

\item{d_vector}{A vector of the d parameters for each cluster}

\item{nbasis}{The number of basis functions}
}
\value{
The updated probability matrix
}
\description{
Update the probabilty matrix
}
\examples{
update_probability_matrix(Y, K, probability_matrix, x, sigma_list, m_list, A_vector, R_vector, d_vector, nbasis)
}
