% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{update_R_vector}
\alias{update_R_vector}
\title{Update the d parameter for each cluster}
\usage{
update_R_vector(
  Y,
  K,
  probability_matrix,
  x,
  sigma_list,
  m_list,
  nbasis,
  beta_vector
)
}
\arguments{
\item{Y}{The matrix containing rows corresponding the curves}

\item{K}{The total number of clusters}

\item{probability_matrix}{A matrix in which the rows represent the probabilities that the curve is in each of the clusters}

\item{x}{The x used to generate the clusters}

\item{sigma_list}{A list of the updated sigma_list parameters for each cluster}

\item{m_list}{A vlist of the updated m parameters for each cluster}

\item{nbasis}{The number of basis functions]}

\item{beta_vector}{A vector of the beta parameters for each cluster}
}
\value{
A vector of the updated R parameters for each cluster
}
\description{
Update the d parameter for each cluster
}
\examples{
update_R_vector(Y, K, probability_matrix, x, sigma_list, m_list, nbasis, beta_vector)

}
