% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{update_sigma_list}
\alias{update_sigma_list}
\title{Update the sigma parameter for each cluster}
\usage{
update_sigma_list(
  Y,
  phi_matrix,
  K,
  A_vector,
  R_vector,
  probability_matrix,
  x,
  nbasis
)
}
\arguments{
\item{Y}{The matrix containing rows corresponding the curves}

\item{phi_matrix}{A matrix of the coffecient vectors for each cluster, phi_k, of the basis matrix B}

\item{K}{The total number of clusters}

\item{A_vector}{A vector}

\item{R_vector}{A vector}

\item{probability_matrix}{A matrix in which the rows represent the probabilities that the curve is in each of the clusters}

\item{x}{The x used to generate the clusters}

\item{nbasis}{The number of basis functions}
}
\value{
A list of the updated sigma parameters for each cluster
}
\description{
Update the sigma parameter for each cluster
}
\examples{
update_sigma_list(Y, K, A_vector, R_vector, probability_matrix, x, nbasis)

}
