% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.R
\name{compute_function}
\alias{compute_function}
\title{Generates simulated with a certain seed, evaluates the data using the model and returns a list with the results from the various evaluation metrics.}
\usage{
compute_function(seed, data_params, model_params, eval_func_list)
}
\arguments{
\item{seed}{The seed}

\item{data_params}{List object containing the parameters required for generating the functional data and its characterics}

\item{model_params}{List object containing the parameters required for model the data and generating the cluster assignments}

\item{eval_func_list}{List object containing the functions corresponding to the various evaluations metrics}
}
\value{
A list with the accuracy metrics as indexes
}
\description{
Generates simulated with a certain seed, evaluates the data using the model and returns a list with the results from the various evaluation metrics.
}
\examples{
compute_function(seed, data_params, model_params, eval_func_list)
}
