% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{funcslustVI}
\alias{funcslustVI}
\title{Generates cluster assignments and related information given functional data.}
\usage{
funcslustVI(
  Y,
  K,
  nbasis,
  x,
  init,
  true_cluster_assignments,
  gamma_dist_config_matrix,
  convergence_threshold,
  verbose
)
}
\arguments{
\item{Y}{A matrix in which the rows represent the curves}

\item{K}{The number of clusters in the data}

\item{nbasis}{The number of basis functions}

\item{x}{The dependent variable}

\item{init}{The initialization method for the algorithim}

\item{true_cluster_assignments}{The true cluster assignments}

\item{gamma_dist_config_matrix}{A matrix where the rows are the alpha and parameters for each cluster}

\item{convergence_threshold}{The threshold that determines when the model has converged}

\item{verbose}{A boolean indicating whether or not to print the inner parameters of the model}
}
\value{
The fitted model
}
\description{
Generates cluster assignments and related information given functional data.
}
\examples{
fit(Y, K, nbasis, x, init, true_cluster_assignments, gamma_dist_config_matrix, convergence_threshold)
}
