% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_v_not_vector}
\alias{get_v_not_vector}
\title{Gets the vnot parameter}
\usage{
get_v_not_vector(Y, x, probability_matrix, phi_matrix, K, nbasis)
}
\arguments{
\item{x}{The x used to generate the clusters}

\item{phi_matrix}{A matrix of the coffecient vectors for each cluster, phi_k, of the basis matrix B}

\item{K}{The total number of clusters}

\item{nbasis}{The number of basis functions}

\item{function_data}{The data generated from the actual functions}
}
\value{
The vnot parameter
}
\description{
Gets the vnot parameter
}
\examples{
get_vnot(x, K)
}
