% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.R
\name{run_simulations}
\alias{run_simulations}
\title{Runs and evaluates the model on the specified simulated function data}
\usage{
run_simulations(
  case,
  x,
  K,
  curves_per_cluster,
  number_of_simulations,
  initialization_method,
  nbasis,
  true_cluster_assignments,
  gamma_dist_config_matrix,
  convergence_threshold,
  save_path,
  verbose,
  draw
)
}
\arguments{
\item{case}{The case used to simulate the data}

\item{x}{The dependent variable}

\item{K}{The number of clusters in the data}

\item{curves_per_cluster}{The number of curves per cluster}

\item{number_of_simulations}{The number of simulations to evatuate the model over}

\item{initialization_method}{The initialization method for the algorithim}

\item{nbasis}{The number of basis functions}

\item{true_cluster_assignments}{The true cluster assignments}

\item{gamma_dist_config_matrix}{A matrix where the rows are the alpha and parameters for each cluster}

\item{convergence_threshold}{The threshold that determines when the model has converged}

\item{save_path}{Pathway to save}

\item{verbose}{A boolean indicating whether or not to print the inner parameters of the model}

\item{draw}{A boolean indicating whether or not to draw the true function vs the}
}
\value{
A matrix with the results of the algorithim for iteration in the simulation for each metric being evaluated over
}
\description{
Runs and evaluates the model on the specified simulated function data
}
\examples{
run_simulations(case, x, K, curves_per_cluster, number_of_simulations, initialization_method, nbasis, true_cluster_assignments, gamma_dist_config_matrix, convergence_threshold, save_path, verbose, draw)
}
