% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{update_R_vector}
\alias{update_R_vector}
\title{Update the d parameter for each cluster}
\usage{
update_R_vector(
  Y,
  K,
  probability_matrix,
  x,
  sigma_list,
  m_list,
  nbasis,
  beta_vector
)
}
\arguments{
\item{Y}{The matrix containing rows corresponding the curves}

\item{K}{The total number of clusters}

\item{probability_matrix}{A matrix in which the rows represent the probabilities that the curve is in each of the clusters}

\item{x}{The x used to generate the clusters}

\item{nbasis}{The number of basis functions}

\item{A_vector}{A vector}

\item{R_vector}{A vector}

\item{sigma_vector}{A vector of the updated sigma parameters for each cluster}

\item{m_vector}{A vector of the updated m parameters for each cluster}
}
\value{
A vector of the updated R parameters for each cluster
}
\description{
Update the d parameter for each cluster
}
\examples{
update_R_vector(K, A_vector, R_vector, probability_matrix, x, nbasis)

}
